Fixed behavior with new threads and caching.
When receiving the alert for a new thread, the value of newValues[op] is undefined, which resulted in the tabCache to receive a NaN item for the thread. 
The localstorage would receive a null value and a reload would just fix everything, but if a connection loss occured the alert sync would push all the updates for that thread into the display, doubling the already displayed amount, since it had no posts it could sum up for that thread.